consolidate attachment preview views
