+++ Mage/UI/Observation/ObservationSummaryViewSwiftUI.swift
@@ -43,89 +43,8 @@ struct ObservationSummaryViewSwiftUI: View {
             if !(viewModel.attachments ?? []).isEmpty {
                 TabView {
                     ForEach(viewModel.orderedAttachments ?? []) { attachment in
-                        if let url = URL(string: attachment.url ?? "") {
-                            if (attachment.contentType?.hasPrefix("image") ?? false) {
-                                KFImage(
-                                    url
-                                )
-                                .requestModifier(ImageCacheProvider.shared.accessTokenModifier)
-                                .cacheOriginalImage()
-                                .onlyFromCache(DataConnectionUtilities.shouldFetchAttachments())
-                                .placeholder {
-                                    Image("observations")
-                                        .symbolRenderingMode(.monochrome)
-                                        .resizable()
-                                        .scaledToFit()
-                                        .foregroundStyle(Color.onSurfaceColor.opacity(0.45))
-                                }
-                                .fade(duration: 0.3)
-                                .resizable()
-                                .scaledToFill()
-                                .frame(maxWidth: .infinity, maxHeight: 150)
-                                .onTapGesture {
-                                    viewModel.appendAttachmentViewRoute(router: router, attachment: attachment)
-                                }
-                            } else if (attachment.contentType?.hasPrefix("video") ?? false),
-                                      let url = attachment.urlWithToken
-                            {
-                                KFImage(
-                                    source: Source.provider(
-                                        AVAssetImageDataProvider(
-                                            assetURL: url,
-                                            time: CMTime(seconds: 0.0, preferredTimescale: 1)
-                                        )
-                                    )
-                                )
-                                .requestModifier(ImageCacheProvider.shared.accessTokenModifier)
-                                .cacheOriginalImage()
-                                .placeholder {
-                                    Image(systemName: "play.circle.fill")
-                                        .symbolRenderingMode(.monochrome)
-                                        .resizable()
-                                        .scaledToFit()
-                                        .foregroundStyle(Color.onSurfaceColor.opacity(0.45))
-                                }
-                            
-                                .fade(duration: 0.3)
-                                .resizable()
-                                .scaledToFill()
-                                .frame(maxWidth: .infinity, maxHeight: 150)
-                                .onTapGesture {
-                                    viewModel.appendAttachmentViewRoute(router: router, attachment: attachment)
-                                }
-                                .overlay {
-                                    Image(systemName: "play.circle.fill")
-                                        .symbolRenderingMode(.monochrome)
-                                        .resizable()
-                                        .scaledToFit()
-                                        .foregroundStyle(Color.onSurfaceColor.opacity(0.45))
-                                        .padding(16)
-                                }
-                            } else if (attachment.contentType?.hasPrefix("audio") ?? false) {
-                                Image(systemName: "speaker.wave.2.fill")
-                                    .symbolRenderingMode(.monochrome)
-                                    .resizable()
-                                    .scaledToFit()
-                                    .foregroundStyle(Color.onSurfaceColor.opacity(0.45))
-                                    .onTapGesture {
-                                        viewModel.appendAttachmentViewRoute(router: router, attachment: attachment)
-                                    }
-                                    .padding(16)
-                                    .frame(maxWidth: .infinity, maxHeight: 150)
-                                    .border(.gray)
-                            } else {
-                                Image(systemName: "paperclip")
-                                    .symbolRenderingMode(.monochrome)
-                                    .resizable()
-                                    .scaledToFit()
-                                    .foregroundStyle(Color.onSurfaceColor.opacity(0.45))
-                                    .onTapGesture {
-                                        viewModel.appendAttachmentViewRoute(router: router, attachment: attachment)
-                                    }
-                                    .padding(16)
-                                    .frame(maxWidth: .infinity, maxHeight: 150)
-                                    .border(.gray)
-                            }
+                        AttachmentPreviewView(attachment: attachment) {
+                            viewModel.appendAttachmentViewRoute(router: router, attachment: attachment)
                         }
                     }
                 }
